Task-3/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
python-docx==1.1.2
PyMuPDF==1.24.9
python-multipart==0.0.9
requests==2.32.3
pydantic-settings==2.4.0
tenacity==8.5.0
langchain-google-genai==1.0.2
langchain-core==0.1.52
pytest==8.3.2
docx2pdf==0.1.8
